Porting to Windows complete
I switched to using timespec_get() because Windows is LLP64 and
its `timeval` is limited to `long` aka int32_t, which is a bummer
as far as measuring the time is concerned.

While there, I removed REMK_PLATFORM_EXTERN because I learned
that you can instruct CMake to just export everything.

While there, I moved stuff in src because looks more similar to
the way in which we're doing things in Measurement Kit.